make NUM_BUILD_THREADS default to 1
this is safer for very slow or memory constrained systems like
aboriginal inside qemu.

also it makes it clear what is really getting built currently
(it happened to users that something that started earlier was still
building, while something started later was already finished,
leading to the impression that the build is stuck when only looking
at the latest logfile).

a poweruser on a fast system can get back the old behaviour
by passing -DNUM_BUILD_THREADS=2 or even higher.
this can speed up builds quite a bit when lots of time is spent
in single-threaded configure scripts, and cores are idle.
context switching seems to be pretty fast on modern cpus, so that
it doesn't hurt much when 2 things are built in parallel with
make -j$MAX_CPUS.