check whether /dev/null exists/is readable on startup
instead of silently failing when posix_spawn is executed.
there used to be some cornercases where /dev/null was missing
(building gcc3 as root caused it to be deleted due to a bug),
and it was completely non-obvious why butch failed and left
only empty log files behind.
it's much more friendly to tell the user what's happening.