outsource responsibility to deal with tarballs to templates
a lot of the complexity in butchs code stems from dealing with
tarball filenames, verification of downloads and so on to pass
easily digestable variables to the dl/build templates.

all of this can be done in the scripts as well, without the need
to recompile butch and publish a new version, which adds a great
deal of complexity.

additionally, the commit introduces handling for a new section
called [vars], which are passed on directly to the script.
this way new variables can be implemented on the fly without
touching butch.
the 3 variables passed via [main], namely filesize, sha512, and
tardir are now handled with the same mechanism used by vars,
and they can be placed in vars or main for backwards compatibility.

the trigger for this change is the naming schemes of the increasing
number of github downloads, which usually name tarballs according
to a release tag like "v0.3.1".
we want to be able to save the tarball under a different name so we
don't get conflicts when 2 projects happen to use the same version
number. also it's much nicer to see "project-foo-v0.3.1.tar.gz" in
the download directory than "v0.3.1.tar.gz" which leaves the user
clueless what it is.
another new issue that arose from gitlab downloads is a URL scheme
with the commit id/tag/branch appended to the tarball name, example
https://gitlab.com/user/prj/repository/archive.tar.bz2?ref=master

if we don't have a method to pass a "save as" filename to the dl
template, we end up with a filename that potentially causes problems
with the shell and additionally doesn't work with the existing
file extension detection mechanisms.

in order for this change to work, a number of changes to the templates
are needed. the following variables are no longer provided:
%BUTCH_TARBALL (can be derived from %BUTCH_MIRROR_URLS)
%BUTCH_HAVE_TARBALL (can be derived from %BUTCH_MIRROR_URLS)
%BUTCH_TARDIR (if specified in pkg, passed as "tardir", if not must
               be derived from tarball name derived from
               %BUTCH_MIRROR_URLS)

the following new variables are provided to templates
(if specified in main or vars section)
tardir
sha512
filesize

as a result, the integrity check via sha512 checksum and/or filesize
must now be done by build and/or dl template.

another sideeffect is that butch now doesn't know anymore if a pkg's
tarball is existing and always launches the download job.
the download template needs to do the checking if the tarball is
existing and return immediately with success if that is the case.

this commits omits the needed update for the built-in templates.
they will be removed in the next commit together with instructions
for where to get compatible templates (link to sabotage linux repo).