enable optional features
it's now possible to request additional features from a package.

consider a package named foo, that can be built with or without
feature bar and feature baz.
however feature bar may cause additional (usually unwanted) bloat,
and feature baz depends on an additional library qux.
so the package foo will by default be built without features bar
and baz.

if you want to install foo with feature bar enabled, you can now
launch "butch install foo.bar".
the package foo.bar may or may not exist, if not existing the only
thing that will be done is the addition of a variable option_bar=1
to the vars section of foo.
this variable can be used to decide in the build section of foo
whether bar shall be enabled or not.
if a package named foo.bar is existing, additionally to setting
option_bar to 1, butch will process its vars and deps sections
and append it to those of foo (this can be used to override vars
from foo).

if you want to install foo with feature baz enabled, there needs
to be a package named foo.baz, which lists qux in its deps section.

it's also possible to "butch install foo.bar.baz" to enable both
bar and baz. the number of possible options is unlimited.

it should be considered good practice to create packages even
if not explicitly needed (as in the case of foo.bar), to document
that the option exists.
please restrict yourself to option names consisting only of alpha-
numeric lowercase characters.

currently this functionality is implemented such that the options
are immediately stripped from the packagename, so if "foo" is
already installed, running "butch install foo.bar" will do nothing.
"butch rebuild foo.bar" otoh will build foo with bar enabled,
however in the butch.db there will be only an entry for "foo".

if a package has foo.bar.baz in its deps, and foo is not already
installed, or was not yet mentioned in another processed package's
deps section, it will be built with bar+baz enabled though.